Refactor QuestionQuiz and WhoKnowsMoreQuiz classes to support dynamic option handling and image/video reveal animations
- Updated QuestionQuiz to handle a variable number of answer options and associated layouts.
- Introduced image loading and resizing functionality for question images.
- Enhanced solution display logic to accommodate both images and videos.
- Modified WhoKnowsMoreQuiz to align with new QuizGameBase constructor parameters.
- Adjusted static file paths for sound assets to ensure compatibility in different environments.
- Removed deprecated translations and files related to the WhoKnowsMoreQuiz.
