+++ launcher.py
@@ -1,4 +1,4 @@
-from buzzinga.buzzinga_class import main
+from buzzinga.src.buzzinga.buzzinga_class import main
 
 if __name__ == "__main__":
     main()
