--- requirements.txt
@@ -1,105 +1,0 @@
-altgraph==0.17.4
-anyio==4.8.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==3.0.0
-async-lru==2.0.4
-attrs==24.3.0
-babel==2.16.0
-beautifulsoup4==4.12.3
-bleach==6.2.0
-certifi==2024.12.14
-cffi==1.17.1
-charset-normalizer==3.4.1
-colorama==0.4.6
-comm==0.2.2
-debugpy==1.8.12
-decorator==5.1.1
-defusedxml==0.7.1
-executing==2.1.0
-fastjsonschema==2.21.1
-fqdn==1.5.1
-h11==0.14.0
-hidapi==0.14.0.post4
-httpcore==1.0.7
-httpx==0.28.1
-idna==3.10
-ipykernel==6.29.5
-ipython==8.31.0
-ipywidgets==8.1.5
-isoduration==20.11.0
-jedi==0.19.2
-Jinja2==3.1.5
-json5==0.10.0
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2024.10.1
-jupyter-console==6.6.3
-jupyter-events==0.11.0
-jupyter-lsp==2.2.5
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-jupyter_server==2.15.0
-jupyter_server_terminals==0.5.3
-jupyterlab==4.3.4
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
-MarkupSafe==3.0.2
-matplotlib-inline==0.1.7
-mistune==3.1.0
-nbclient==0.10.2
-nbconvert==7.16.5
-nbformat==5.10.4
-nest-asyncio==1.6.0
-notebook==7.3.2
-notebook_shim==0.2.4
-overrides==7.7.0
-packaging==24.2
-pandocfilters==1.5.1
-parso==0.8.4
-pefile==2023.2.7
-pillow==11.1.0
-platformdirs==4.3.6
-prometheus_client==0.21.1
-prompt_toolkit==3.0.48
-psutil==6.1.1
-pure_eval==0.2.3
-pybuzzers==1.0.6
-pycparser==2.22
-pygame==2.6.1
-Pygments==2.19.1
-pyinstaller==6.11.1
-pyinstaller-hooks-contrib==2025.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-python-json-logger==3.2.1
-pywin32==308
-pywin32-ctypes==0.2.3
-pywinpty==2.0.14
-PyYAML==6.0.2
-pyzmq==26.2.0
-referencing==0.36.1
-requests==2.32.3
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rpds-py==0.22.3
-Send2Trash==1.8.3
-setuptools==75.8.0
-six==1.17.0
-sniffio==1.3.1
-soupsieve==2.6
-stack-data==0.6.3
-terminado==0.18.1
-tinycss2==1.4.0
-tornado==6.4.2
-traitlets==5.14.3
-types-python-dateutil==2.9.0.20241206
-uri-template==1.3.0
-urllib3==2.3.0
-wcwidth==0.2.13
-webcolors==24.11.1
-webencodings==0.5.1
-websocket-client==1.8.0
-widgetsnbextension==4.0.13
