+++ src/buzzinga/animation.py
@@ -1,4 +1,4 @@
-import pygame
+import pygame, os
 from .static import Static
 from .game_utilities import load_and_scale_image
 
@@ -28,13 +28,13 @@ class BuzzingaAnimation(Animation):
         super().__init__()
         self.is_animating = False
         self.sprites = []
-        self.sprites.append(load_and_scale_image('Buzzinga-1.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
-        self.sprites.append(load_and_scale_image('Buzzinga-2.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
-        self.sprites.append(load_and_scale_image('Buzzinga-3.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
-        self.sprites.append(load_and_scale_image('Buzzinga-4.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
-        self.sprites.append(load_and_scale_image('Buzzinga-5.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
-        self.sprites.append(load_and_scale_image('Buzzinga-6.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
-        self.sprites.append(load_and_scale_image('Buzzinga-7.png', Static.STATIC_FOLDER, rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-1.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-2.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-3.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-4.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-5.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-6.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
+        self.sprites.append(load_and_scale_image('Buzzinga-7.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.w / 2, image_cache))
         self.current_sprite = 0
         self.image = self.sprites[self.current_sprite]
 
@@ -47,16 +47,16 @@ class SoundAnimation(Animation):
         super().__init__()
         self.is_animating = False
         self.sprites = []
-        self.sprites.append(load_and_scale_image('sound_animation-1.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-2.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-3.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-4.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-5.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-6.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-7.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-8.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-9.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
-        self.sprites.append(load_and_scale_image('sound_animation-10.png', Static.STATIC_FOLDER, rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-1.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-2.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-3.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-4.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-5.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-6.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-7.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-8.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-9.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
+        self.sprites.append(load_and_scale_image('sound_animation-10.png', os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER), rect.h*0.7, image_cache))
         self.current_sprite = 0
         self.image = self.sprites[self.current_sprite]
