--- translations.py
@@ -1,95 +1,0 @@
-
-english = {
-    'progress': 'progress',
-    'correct': 'correct',
-    'wrong': 'wrong',
-    'next_hint': 'next hint',
-    'display_answer': 'display answer',
-    'player1': 'Player 1',
-    'player2': 'Player 2',
-    'player3': 'Player 3',
-    'player4': 'Player 4',
-    'deletion_successful': 'Deletion successful',
-    'deletion_failed': 'Deletion failed',
-    'press_any_key': 'PRESS ANY KEY',
-    'who_plays': 'W H O  P L A Y S ?',
-    'language': 'LANGUAGE',
-    'choose_game': 'Choose Game',
-    'which_game': 'W H I C H  G A M E ?',
-    'image_quiz': 'Image Quiz',
-    'audio_quiz': 'Audio Quiz',
-    'hints': '10 Hints',
-    'who_knows_more': 'Who knows more?',
-    'back': 'Back',
-    'game_category': 'W H I C H  C A T E G O R Y ?',
-    'free_disk_space': 'free disk space',
-    'no_categories': 'No categories exist!',
-    'settings': 'S E T T I N G S',
-    'mode': 'MODE',
-    'no_buzzer': 'No Buzzer connected!',
-    'on': 'On',
-    'off': 'Off',
-    'all_rounds': 'All rounds',
-    'points': 'Points',
-    'start_game': 'Start Game',
-    'keys': 'Keys',
-    'import_files': 'Import Files',
-    'successful_import': 'Files imported successfully',
-    'failed_import': 'Error during file import',
-    'no_files_imported': 'No files imported',
-    'rounds': 'Rounds',
-    'round': 'Round',
-    'winner': 'WINNER',
-    'already_solved': 'Already solved!',
-    'all_solved': 'All solved!',
-    'wrong_id': 'Wrong ID',
-    'reverse_wrong_input': 'Reverse wrong input'
-
-}
-
-german = {
-    'progress': 'weiter',
-    'correct': 'richtig',
-    'wrong': 'falsch',
-    'next_hint': 'nächster Hinweis',
-    'display_answer': 'Antwort einblenden',
-    'player1': 'Spieler 1',
-    'player2': 'Spieler 2',
-    'player3': 'Spieler 3',
-    'player4': 'Spieler 4',
-    'deletion_successful': 'Löschen erfolgreich',
-    'deletion_failed': 'Löschen fehlgeschlagen',
-    'who_plays': 'W E R  S P I E L T  M I T ?',
-    'language': 'SPRACHE',
-    'choose_game': 'Spiel auswählen',
-    'which_game': 'W E L C H E S  S P I E L ?',
-    'image_quiz': 'Bilder Quiz',
-    'hints': '10 Hinweise',
-    'who_knows_more': 'Wer weiß mehr?',
-    'back': 'Zurück',
-    'game_category': 'W E L C H E  K A T E G O R I E ?',
-    'free_disk_space': 'freier Speicherplatz',
-    'no_categories': 'Es existieren keine Kategorien!',
-    'settings': 'E I N S T E L L U N G E N',
-    'mode': 'MODUS',
-    'no_buzzer': 'Kein Buzzer verbunden!',
-    'on': 'An',
-    'off': 'Aus',
-    'all_rounds': 'Alle Runden',
-    'points': 'Punkte',
-    'start_game': 'Spiel starten',
-    'keys': 'Tasten',
-    'import_files': 'Importiere Dateien',
-    'successful_import': 'Dateien erfolgreich importiert',
-    'failed_import': 'Fehler beim Import der Dateien',
-    'no_files_imported': 'Keine Dateien importiert',
-    'rounds': 'Runden',
-    'round': 'Runde',
-    'winner': 'GEWINNER',
-    'already_solved': 'Schon gelöst!',
-    'all_solved': 'Alles gelöst!',
-    'wrong_id': 'Falsche ID',
-    'reverse_wrong_input': 'Korrektur',
-    'reveal': 'aufdecken',
-    'reveal animation': 'Aufdeckanimation'
-}
